chore: move facility designated operation timeline schemas to v2
